Split into private and public selectors/actions
