+++ packages/upload-media/src/store/test/reducer.ts
@@ -11,12 +11,12 @@ describe( 'reducer', () => {
 	describe( `${ Type.Add }`, () => {
 		it( 'adds an item to the queue', () => {
 			const initialState: State = {
-				imageSizes: {},
 				queueStatus: 'active',
 				blobUrls: {},
 				settings: {
 					mediaUpload: jest.fn(),
 					mediaSideload: jest.fn(),
+					imageSizes: {},
 				},
 				queue: [
 					{
@@ -34,12 +34,12 @@ describe( 'reducer', () => {
 			} );
 
 			expect( state ).toEqual( {
-				imageSizes: {},
 				queueStatus: 'active',
 				blobUrls: {},
 				settings: {
 					mediaUpload: expect.any( Function ),
 					mediaSideload: expect.any( Function ),
+					imageSizes: {},
 				},
 				queue: [
 					{
@@ -58,12 +58,12 @@ describe( 'reducer', () => {
 	describe( `${ Type.Cancel }`, () => {
 		it( 'removes an item from the queue', () => {
 			const initialState: State = {
-				imageSizes: {},
 				queueStatus: 'active',
 				blobUrls: {},
 				settings: {
 					mediaUpload: jest.fn(),
 					mediaSideload: jest.fn(),
+					imageSizes: {},
 				},
 				queue: [
 					{
@@ -83,12 +83,12 @@ describe( 'reducer', () => {
 			} );
 
 			expect( state ).toEqual( {
-				imageSizes: {},
 				queueStatus: 'active',
 				blobUrls: {},
 				settings: {
 					mediaUpload: expect.any( Function ),
 					mediaSideload: expect.any( Function ),
+					imageSizes: {},
 				},
 				queue: [
 					{
@@ -108,12 +108,12 @@ describe( 'reducer', () => {
 	describe( `${ Type.Remove }`, () => {
 		it( 'removes an item from the queue', () => {
 			const initialState: State = {
-				imageSizes: {},
 				queueStatus: 'active',
 				blobUrls: {},
 				settings: {
 					mediaUpload: jest.fn(),
 					mediaSideload: jest.fn(),
+					imageSizes: {},
 				},
 				queue: [
 					{
@@ -132,12 +132,12 @@ describe( 'reducer', () => {
 			} );
 
 			expect( state ).toEqual( {
-				imageSizes: {},
 				queueStatus: 'active',
 				blobUrls: {},
 				settings: {
 					mediaUpload: expect.any( Function ),
 					mediaSideload: expect.any( Function ),
+					imageSizes: {},
 				},
 				queue: [
 					{
@@ -152,12 +152,12 @@ describe( 'reducer', () => {
 	describe( `${ Type.PauseItem }`, () => {
 		it( 'marks an item as paused', () => {
 			const initialState: State = {
-				imageSizes: {},
 				queueStatus: 'active',
 				blobUrls: {},
 				settings: {
 					mediaUpload: jest.fn(),
 					mediaSideload: jest.fn(),
+					imageSizes: {},
 				},
 				queue: [
 					{
@@ -176,12 +176,12 @@ describe( 'reducer', () => {
 			} );
 
 			expect( state ).toEqual( {
-				imageSizes: {},
 				queueStatus: 'active',
 				blobUrls: {},
 				settings: {
 					mediaUpload: expect.any( Function ),
 					mediaSideload: expect.any( Function ),
+					imageSizes: {},
 				},
 				queue: [
 					{
@@ -200,12 +200,12 @@ describe( 'reducer', () => {
 	describe( `${ Type.ResumeItem }`, () => {
 		it( 'marks an item as processing', () => {
 			const initialState: State = {
-				imageSizes: {},
 				queueStatus: 'active',
 				blobUrls: {},
 				settings: {
 					mediaUpload: jest.fn(),
 					mediaSideload: jest.fn(),
+					imageSizes: {},
 				},
 				queue: [
 					{
@@ -224,12 +224,12 @@ describe( 'reducer', () => {
 			} );
 
 			expect( state ).toEqual( {
-				imageSizes: {},
 				queueStatus: 'active',
 				blobUrls: {},
 				settings: {
 					mediaUpload: expect.any( Function ),
 					mediaSideload: expect.any( Function ),
+					imageSizes: {},
 				},
 				queue: [
 					{
@@ -248,12 +248,12 @@ describe( 'reducer', () => {
 	describe( `${ Type.AddOperations }`, () => {
 		it( 'appends operations to the list', () => {
 			const initialState: State = {
-				imageSizes: {},
 				queueStatus: 'active',
 				blobUrls: {},
 				settings: {
 					mediaUpload: jest.fn(),
 					mediaSideload: jest.fn(),
+					imageSizes: {},
 				},
 				queue: [
 					{
@@ -270,12 +270,12 @@ describe( 'reducer', () => {
 			} );
 
 			expect( state ).toEqual( {
-				imageSizes: {},
 				queueStatus: 'active',
 				blobUrls: {},
 				settings: {
 					mediaUpload: expect.any( Function ),
 					mediaSideload: expect.any( Function ),
+					imageSizes: {},
 				},
 				queue: [
 					{
@@ -295,12 +295,12 @@ describe( 'reducer', () => {
 	describe( `${ Type.OperationStart }`, () => {
 		it( 'marks an item as processing', () => {
 			const initialState: State = {
-				imageSizes: {},
 				queueStatus: 'active',
 				blobUrls: {},
 				settings: {
 					mediaUpload: jest.fn(),
 					mediaSideload: jest.fn(),
+					imageSizes: {},
 				},
 				queue: [
 					{
@@ -328,12 +328,12 @@ describe( 'reducer', () => {
 			} );
 
 			expect( state ).toEqual( {
-				imageSizes: {},
 				queueStatus: 'active',
 				blobUrls: {},
 				settings: {
 					mediaUpload: expect.any( Function ),
 					mediaSideload: expect.any( Function ),
+					imageSizes: {},
 				},
 				queue: [
 					{
@@ -361,12 +361,12 @@ describe( 'reducer', () => {
 	describe( `${ Type.OperationFinish }`, () => {
 		it( 'marks an item as processing', () => {
 			const initialState: State = {
-				imageSizes: {},
 				queueStatus: 'active',
 				blobUrls: {},
 				settings: {
 					mediaUpload: jest.fn(),
 					mediaSideload: jest.fn(),
+					imageSizes: {},
 				},
 				queue: [
 					{
@@ -389,12 +389,12 @@ describe( 'reducer', () => {
 			} );
 
 			expect( state ).toEqual( {
-				imageSizes: {},
 				queueStatus: 'active',
 				blobUrls: {},
 				settings: {
 					mediaUpload: expect.any( Function ),
 					mediaSideload: expect.any( Function ),
+					imageSizes: {},
 				},
 				queue: [
 					{
